T#769: Extract risk routes to src/risk/routes.ts
9 handlers (risk CRUD + summary + stale + risk comments) moved from
server.ts to src/risk/routes.ts via registerRiskRoutes DI pattern.
DDL for risks + risk_comments tables stays in server.ts.

server.ts: 12,598 → 12,335 lines (-263). No logic changes.
